// src/types/search-result.interface.ts
export interface SearchResult {
  supplier: string;
  service: string;
  period_from: string;
  period_to: string;
  pax_from: number;
  pax_to: number;
  price: number;
  price_for: string;
  days: string;
  market: string;
  category: string;
  currency: string;
  active: string;
  estimated: string;
}